+++ src/api/client.ts
@@ -3,7 +3,7 @@
 *   Copyright © 2024 NatML Inc. All Rights Reserved.
 */
 
-import { FunctionConfig } from "../function"
+import type { FunctionConfig } from "../function"
 
 export interface GraphPayload<T> {
     /**